+++ src/xrGame/death_anims.cpp
@@ -36,7 +36,7 @@ MotionID	rnd_motion::	motion		( )	const
 {
 	if( motions.empty( ) )
 			return MotionID();
-	return motions[ ::Random.randI(0, motions.size()) ];
+	return motions[ ::Random.randI(0, (u32)motions.size()) ];
 }
