+++ src/xrServerEntities/xml_str_id_loader.h
@@ -64,7 +64,7 @@ public:
 		return item?item->id:default_id;
 	}
 
-	static const int		GetMaxIndex	()					 {return m_pItemDataVector->size()-1;}
+	static const int		GetMaxIndex	()					 {return (int)m_pItemDataVector->size()-1;}
 
 	//удаление статичекого массива
 	static void					DeleteIdToIndexData		();
